refactor(ui): move the popover client component from primitives to components
