+++ packages/form/resources/js/rich-editor/builtins.tsx
@@ -14,7 +14,11 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@lattice-php/ui/primitives/dropdown-menu";
-import { Popover, PopoverContent, PopoverTrigger } from "@lattice-php/ui/primitives/popover";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@lattice-php/ui/components/popover/popover";
 import type { RichEditorExtensionRegistryFor, ToolbarButton } from "./registry";
 import { ToolbarIconButton } from "./toolbar-button";
