+++ packages/table/resources/js/components/filter-bar.tsx
@@ -1,5 +1,9 @@
 import { IconButton } from "@lattice-php/ui/primitives/icon-button";
-import { Popover, PopoverContent, PopoverTrigger } from "@lattice-php/ui/primitives/popover";
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from "@lattice-php/ui/components/popover/popover";
 import { useT } from "@lattice-php/ui/i18n";
 import type { Option } from "@lattice-php/core";
 import { isActiveFilterValue } from "@lattice-php/table/lib/filter-values";
